requirements.txt
# API clients
requests==2.31.0
openai==1.54.0

# Data processing
python-dotenv==1.0.0
pydantic==2.4.2
pyyaml==6.0.1
psutil==5.9.6

# Temporal
temporalio==1.5.0

# CLI
click==8.1.7
rich==13.6.0  # For beautiful CLI output

# Date/time handling
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.1  # For mocking API calls

